requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
torch>=2.8.0
sentence-transformers>=5.1.0
faiss-cpu>=1.7.4
neo4j>=5.20.0
langchain>=0.2.0
langchain-community>=0.2.0
google-generativeai>=0.8.5
pypdf2>=3.0.1
python-multipart>=0.0.7
pydantic>=2.7.0
transformers>=4.42.0
pandas>=2.0.0